+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Install DNS lookup utilities (`dig`, `host`, `nslookup`) in the `cli` profile.
 - Install iproute2 (`ip`) in the `cli` profile.
 - Install ping in the `cli` profile.
+- Install rsync in the `cli` profile.
 
 ## [1.6.0] - 2026-08-10